integrated_system/requirements.txt
# 车道检测系统集成版依赖
# Integrated Lane Detection System Dependencies

# 核心依赖
numpy>=1.19.0
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0

# 深度学习
tensorrt>=8.0.0
cuda-python>=12.0.0

# 串口通信
pyserial>=3.5

# 图像处理
Pillow>=8.0.0

# 日志
colorlog>=6.0.0

# 可选依赖（用于开发和测试）
# pytest>=6.0.0
# matplotlib>=3.3.0